[Fix] Fix warning in 669 loader
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4923 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_669.cpp
@@ -209,7 +209,7 @@ bool CSoundFile::Read669(FileReader &file, ModLoadingFlags loadFlags)
 
 				m->param = effect[chn] & 0x0F;
 
-				if((effect[chn] >> 4) < CountOf(effTrans))
+				if((effect[chn] >> 4) < static_cast<uint8>(CountOf(effTrans)))
 				{
 					m->command = effTrans[effect[chn] >> 4];
 				} else
